backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
SQLAlchemy>=2.0,<3.0
aiomysql>=0.2,<1.0
pydantic[email]>=2.10,<3.0
pydantic-settings>=2.8,<3.0
python-jose[cryptography]>=3.3,<4.0
passlib[argon2,bcrypt]>=1.7,<2.0
python-multipart>=0.0.20,<1.0
pyotp>=2.9,<3.0
qrcode[pil]>=8.0,<9.0
cryptography>=46.0,<47.0
Pillow>=11.0,<12.0
httpx>=0.28,<1.0
reportlab>=4.2,<5.0
aiosmtplib>=3.0,<4.0
numpy>=2.0,<3.0
opencv-python-headless>=4.10,<5.0
scikit-learn>=1.6,<2.0
